remove old z-scale spherical harmonics, replace with scaling along principle axes
